refactor: manager 책임 분리 및 결과 책임 추가
